refactor : adjust match id and money type to long from int

+++ lib_database/src/main/java/arch/cayenne/lib/database/GameTypeConverter.kt
@@ -18,7 +18,7 @@ class GameTypeConverter {
         return value.split(",").let {
             Selection(
                 marketName = it[0],
-                id = it[1].toInt(),
+                id = it[1].toLong(),
                 name = it[2],
                 odds = it[3].toInt()
             )
