+++ lib_database/src/main/java/arch/cayenne/lib/database/entity/BetBean.kt
@@ -19,15 +19,15 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "BetBean")
 data class BetBean(
     @PrimaryKey
-    val matchId: Int, // 赛事ID
+    val matchId: Long, // 赛事ID
     var selection: Selection, // 选择的盘口
     var reverseOdds: Int? = null, // 預約赔率
     var betType: BetTypeEnum, // 0: 單注 1: 串關 2: 預約
     var status: BetStatusEnum = BetStatusEnum.PENDING_BET, // 下注狀態
     val leagueName: String, // 联赛名称 ex. 世界盃
     val matchName: String, // 赛事名称 ex. 中國 vs 日本
-    var minAmount: Int, // 最小下注金额
-    var maxAmount: Int, // 最大下注金额
+    var minAmount: Long, // 最小下注金额
+    var maxAmount: Long, // 最大下注金额
     var isBetStop: Boolean = false, // 是否停止下注
     val isPlaying: Boolean = false, // 是否滾球
 )
@@ -40,7 +40,7 @@ data class BetBean(
  */
 data class Selection(
     val marketName: String, // 盘口名称 ex. 讓分盤
-    val id: Int, // 盘口ID
+    val id: Long, // 盘口ID
     var name: String, // 盘口名称 ex. 中國 (+1.5)
     var odds: Int // 盘口赔率 ex. 1.9
 )
